requirements/dev.txt
-r common.txt
pytest-django>=2.9.1
coverage>=4.0.3
